pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "yara-compendium"
description = "Yara rule aggregation and management tool"
version = "0.3.0"
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "GitPython",
    "plyara~=2.1.1",
    "PyYAML",
#    "yara-python~=4.2.0",  # Note: don't install yara-python via pip so it can be properly linked
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
yara-compendium = "compendium.main:main"

[project.optional-dependencies]
tests = [
    "black",
    "flake8",
    "isort",
    "pylint~=2.14.5",
    "pytest~=7.1.2",
    "pytest-cov~=3.0.0",
    "tox",
]

[tool.setuptools.packages.find]
include = ["compendium"]

[tool.setuptools.package-data]
etc = [
    "etc/config.yml",
    "etc/disable.conf",
    "etc/modify.conf",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
)/
'''
